monitor: Eliminate a redundant failable allocation
Using g_try_new() here is incorrect: if it fails, it will return NULL
but the code will proceed to set the array elements, which will result
in an invalid write and a crash.

Since this method is called just after lots of allocations for loading
the module, assume it’s going to succeed (just like we do with the rest
of the allocations ever).