monitor: Add dummy support for the GNetworkMonitor:connectivity property
This is needed to prevent warnings about not implementing a required
property when constructing a GNetworkMonitorConnman. The current
implementation always assumes no captive portals when connected, and
does not consult ConnMan to check this.

A better future implementation would check the relevant ConnMan
properties, but that is beyond my capability to test at the moment.